Extract generation of minion pillar files
