+++ java/code/src/com/suse/manager/webui/services/SaltStateGeneratorService.java
@@ -15,38 +15,29 @@
 package com.suse.manager.webui.services;
 
 import static com.suse.manager.webui.services.SaltConstants.PILLAR_DATA_FILE_EXT;
-import static com.suse.manager.webui.services.SaltConstants.PILLAR_DATA_FILE_PREFIX;
 import static com.suse.manager.webui.services.SaltConstants.PILLAR_IMAGE_DATA_FILE_EXT;
 import static com.suse.manager.webui.services.SaltConstants.PILLAR_IMAGE_DATA_FILE_PREFIX;
 import static com.suse.manager.webui.services.SaltConstants.SALT_CONFIG_STATES_DIR;
 import static com.suse.manager.webui.services.SaltConstants.SALT_SERVER_STATE_FILE_PREFIX;
-import static com.suse.manager.webui.services.SaltConstants.SUMA_PILLAR_DATA_PATH;
 import static com.suse.manager.webui.services.SaltConstants.SUMA_PILLAR_IMAGES_DATA_PATH;
 import static com.suse.manager.webui.services.SaltConstants.SUMA_STATE_FILES_ROOT_PATH;
 import static com.suse.manager.webui.utils.SaltFileUtils.defaultExtension;
 
-import com.redhat.rhn.common.conf.Config;
 import com.redhat.rhn.common.conf.ConfigDefaults;
 import com.redhat.rhn.common.util.FileUtils;
-import com.redhat.rhn.domain.channel.AccessToken;
-import com.redhat.rhn.domain.channel.AccessTokenFactory;
-import com.redhat.rhn.domain.channel.Channel;
 import com.redhat.rhn.domain.config.ConfigChannel;
 import com.redhat.rhn.domain.org.Org;
 import com.redhat.rhn.domain.server.MinionServer;
 import com.redhat.rhn.domain.server.MinionServerFactory;
-import com.redhat.rhn.domain.server.Server;
 import com.redhat.rhn.domain.server.ServerGroup;
-import com.redhat.rhn.domain.server.ServerGroupType;
-import com.redhat.rhn.domain.server.ServerPath;
 import com.redhat.rhn.domain.state.OrgStateRevision;
 import com.redhat.rhn.domain.state.ServerGroupStateRevision;
 import com.redhat.rhn.domain.state.ServerStateRevision;
 import com.redhat.rhn.domain.state.StateFactory;
 import com.redhat.rhn.domain.state.StateRevision;
 import com.redhat.rhn.domain.user.User;
-import com.suse.manager.utils.MachinePasswordUtils;
 import com.suse.manager.webui.controllers.StatesAPI;
+import com.suse.manager.webui.services.pillar.MinionPillarManager;
 import com.suse.manager.webui.utils.SaltConfigChannelState;
 import com.suse.manager.webui.utils.SaltPillar;
 import com.suse.manager.webui.utils.SaltStateGenerator;
@@ -59,10 +50,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -77,145 +65,13 @@ public enum SaltStateGeneratorService {
     // Singleton instance of this class
     INSTANCE;
 
-    private static final Map<String, Object> PKGSET_BEACON_PROPS = new HashMap<>();
-    private static final Map<String, Object> VIRTPOLLER_BEACON_PROPS = new HashMap<>();
-
-    private static final String PKGSET_COOKIE_PATH = "/var/cache/salt/minion/rpmdb.cookie";
-    private static final int PKGSET_INTERVAL = 5;
-    static {
-        PKGSET_BEACON_PROPS.put("cookie", PKGSET_COOKIE_PATH);
-        PKGSET_BEACON_PROPS.put("interval",  PKGSET_INTERVAL);
-    }
-
-    static {
-        VIRTPOLLER_BEACON_PROPS.put("cache_file", Config.get().getString(
-                ConfigDefaults.VIRTPOLLER_CACHE_FILE));
-        VIRTPOLLER_BEACON_PROPS.put("expire_time", Config.get().getInt(
-                ConfigDefaults.VIRTPOLLER_CACHE_EXPIRATION));
-        VIRTPOLLER_BEACON_PROPS.put("interval", Config.get().getInt(
-                ConfigDefaults.VIRTPOLLER_INTERVAL));
-    }
-
     /** Logger */
     private static final Logger LOG = Logger.getLogger(SaltStateGeneratorService.class);
 
     private Path suseManagerStatesFilesRoot;
 
-    private Path pillarDataPath;
-
     SaltStateGeneratorService() {
         suseManagerStatesFilesRoot = Paths.get(SUMA_STATE_FILES_ROOT_PATH);
-        pillarDataPath = Paths.get(SUMA_PILLAR_DATA_PATH);
-    }
-
-    /**
-     * Generate server specific pillar if the given server is a minion.
-     * @param minion the minion server
-     */
-    public void generatePillar(MinionServer minion) {
-        generatePillar(minion, true, Collections.emptySet());
-    }
-
-    /**
-     * Creates the directory to store the pillar file on disk.
-     * @throws IOException if an I/O error occurs
-     */
-    public void createPillarDirectory() throws IOException {
-        Files.createDirectories(pillarDataPath);
-    }
-
-    /**
-     * Generate pillar containing the information of the server groups the the passed minion is member of
-     * @param minion the minion server
-     */
-    public void generateGroupMemebershipsPillarData(MinionServer minion) {
-        LOG.debug("Generating group memberships pillar file for minion: " + minion.getMinionId());
-
-        SaltPillar pillar = new SaltPillar();
-
-        List<String> addonGroupTypes = minion.getEntitledGroupTypes().stream()
-                .map(ServerGroupType::getLabel).collect(Collectors.toList());
-
-        List<Long> groupIds = minion.getManagedGroups().stream().map(ServerGroup::getId).collect(Collectors.toList());
-
-        pillar.add("group_ids", groupIds.toArray(new Long[groupIds.size()]));
-        pillar.add("addon_group_types", addonGroupTypes.toArray(new String[addonGroupTypes.size()]));
-
-        try {
-            Files.createDirectories(pillarDataPath);
-            Path filePath = pillarDataPath.resolve(getServerGroupMembershipPillarFileName(minion.getMinionId()));
-            new SaltStateGenerator(filePath.toFile()).generate(pillar);
-        }
-        catch (IOException e) {
-            LOG.error(e.getMessage(), e);
-        }
-    }
-
-    /**
-     * Generate server specific pillar if the given server is a minion.
-     * @param minion the minion server
-     * @param refreshAccessTokens if access tokens should be refreshed first
-     * @param tokensToActivate channels access tokens to activate when refreshing
-     * the tokens
-     */
-    public void generatePillar(MinionServer minion, boolean refreshAccessTokens,
-                               Collection<AccessToken> tokensToActivate) {
-        LOG.debug("Generating pillar file for minion: " + minion.getMinionId());
-
-        generateGroupMemebershipsPillarData(minion);
-
-        if (refreshAccessTokens) {
-            AccessTokenFactory.refreshTokens(minion, tokensToActivate);
-        }
-
-        SaltPillar pillar = new SaltPillar();
-        pillar.add("org_id", minion.getOrg().getId());
-
-        pillar.add("contact_method", minion.getContactMethod().getLabel());
-        pillar.add("mgr_server", getChannelHost(minion));
-        pillar.add("machine_password", MachinePasswordUtils.machinePassword(minion));
-
-        Map<String, Object> chanPillar = new HashMap<>();
-        minion.getAccessTokens().stream().filter(AccessToken::getValid).forEach(accessToken -> {
-            accessToken.getChannels().forEach(chan -> {
-                Map<String, Object> chanProps = getChannelPillarData(minion, accessToken, chan);
-
-                chanPillar.put(chan.getLabel(), chanProps);
-            });
-        });
-        pillar.add("channels", chanPillar);
-
-        Map<String, Object> beaconConfig = new HashMap<>();
-        // this add the configuration for the beacon that tell us when the
-        // minion packages are modified locally
-        if (minion.getOsFamily().toLowerCase().equals("suse") ||
-                minion.getOsFamily().toLowerCase().equals("redhat")) {
-            beaconConfig.put("pkgset", PKGSET_BEACON_PROPS);
-        }
-        // this add the configuration for the beacon that tell us about
-        // virtual guests running on that minion
-        // The virtpoller is still usefull with the libvirt events: it will help
-        // synchronizing the DB with the actual guest lists in case we had a temporary shutdown.
-        // TODO: find a better way to detect when the beacon should be configured
-        if (minion.isVirtualHost()) {
-            beaconConfig.put("virtpoller", VIRTPOLLER_BEACON_PROPS);
-        }
-        if (!beaconConfig.isEmpty()) {
-            pillar.add("beacons", beaconConfig);
-        }
-
-        try {
-            Files.createDirectories(pillarDataPath);
-            Path filePath = pillarDataPath.resolve(
-                    getServerPillarFileName(minion.getMinionId())
-            );
-            com.suse.manager.webui.utils.SaltStateGenerator saltStateGenerator =
-                    new com.suse.manager.webui.utils.SaltStateGenerator(filePath.toFile());
-            saltStateGenerator.generate(pillar);
-        }
-        catch (IOException e) {
-            LOG.error(e.getMessage(), e);
-        }
     }
 
     /**
@@ -329,114 +185,10 @@ public enum SaltStateGeneratorService {
         return bootImagePillar;
     }
 
-    /**
-     * Create channel pillar data for the given minion, access token and channel.
-     * @param minion the minion
-     * @param accessToken the access token
-     * @param chan the channel
-     * @return a {@link Map} containing the pillar data
-     */
-    public static Map<String, Object> getChannelPillarData(MinionServer minion, AccessToken accessToken, Channel chan) {
-        Map<String, Object> chanProps = new HashMap<>();
-        chanProps.put("alias", "susemanager:" + chan.getLabel());
-        chanProps.put("name", chan.getName());
-        chanProps.put("enabled", "1");
-        chanProps.put("autorefresh", "1");
-        chanProps.put("host", getChannelHost(minion));
-        if ("ssh-push-tunnel".equals(minion.getContactMethod().getLabel())) {
-            chanProps.put("port", Config.get().getInt("ssh_push_port_https"));
-        }
-        chanProps.put("token", accessToken.getToken());
-        if (chan.isTypeRpm()) {
-            chanProps.put("type", "rpm-md");
-        }
-        else if (chan.isTypeDeb()) {
-            chanProps.put("type", "deb");
-        }
-        else {
-            LOG.warn("Unknown repo type for channel " + chan.getLabel());
-        }
-
-        if (ConfigDefaults.get().isMetadataSigningEnabled()) {
-            chanProps.put("gpgcheck", chan.isGPGCheck() ? "1" : "0");
-            // three state field. yes, no or default
-            chanProps.put("repo_gpgcheck", "default");
-            chanProps.put("pkg_gpgcheck", "default");
-        }
-        else {
-            chanProps.put("gpgcheck", "0");
-            chanProps.put("repo_gpgcheck", "0");
-            chanProps.put("pkg_gpgcheck", chan.isGPGCheck() ? "1" : "0");
-        }
-        return chanProps;
-    }
-
-    /**
-     * Return the channel hostname for a given server.
-     *
-     * @param server server to get the channel host for.
-     * @return channel hostname.
-     */
-    public static String getChannelHost(Server server) {
-        Optional<ServerPath> path = server.getFirstServerPath();
-        if (!path.isPresent()) {
-            // client is not proxied, return this server's hostname
-            // HACK: we currently have no better way
-            return ConfigDefaults.get().getCobblerHost();
-        }
-        else {
-            return path.get().getHostname();
-        }
-    }
-
-    private String getServerGroupMembershipPillarFileName(String minionId) {
-        return PILLAR_DATA_FILE_PREFIX + "_" +
-                minionId + "_" + "group_memberships" + "." +
-                PILLAR_DATA_FILE_EXT;
-    }
-
-    private String getServerPillarFileName(String minionId) {
-        return PILLAR_DATA_FILE_PREFIX + "_" +
-            minionId + "." +
-                PILLAR_DATA_FILE_EXT;
-    }
-
     private String getImagePillarFileName(OSImageInspectSlsResult.Bundle bundle) {
         return PILLAR_IMAGE_DATA_FILE_PREFIX + "-" + bundle.getBasename() + "-" +
                 bundle.getId().replace('.', '-') + "." + PILLAR_IMAGE_DATA_FILE_EXT;
     }
-
-    /**
-     * Remove the corresponding pillar data of minion server
-     * @param minion the minion server
-     */
-    public void removePillar(MinionServer minion) {
-        removePillar(minion.getMinionId());
-    }
-
-    /**
-     * Remove the corresponding pillar data by minionId
-     * @param minionId the server minionId
-     */
-    public void removePillar(String minionId) {
-        LOG.debug("Removing pillar file for minion: " + minionId);
-        this.getMinionPillarFilenames(minionId).stream().forEach(filename -> removePillarFile(filename));
-    }
-
-    private List<String> getMinionPillarFilenames(String minionId) {
-        return Arrays.asList(getServerPillarFileName(minionId), getServerGroupMembershipPillarFileName(minionId));
-    }
-
-    private void removePillarFile(String filename) {
-        Path filePath = pillarDataPath.resolve(filename);
-        try {
-            Files.deleteIfExists(filePath);
-        }
-        catch (IOException e) {
-            LOG.error("Could not remove pillar file " + filePath);
-        }
-    }
-
 
     /**
      * Remove the config channel assignments for minion server.
@@ -570,7 +322,7 @@ public enum SaltStateGeneratorService {
      * @param minion the minion
      */
     public void removeServer(MinionServer minion) {
-        removePillar(minion);
+        MinionPillarManager.INSTANCE.removePillar(minion.getMinionId());
         removeConfigChannelAssignments(minion);
         removeActionChains(minion);
     }
@@ -627,7 +379,7 @@ public enum SaltStateGeneratorService {
         StateFactory.save(newStateRev);
 
         // refresh pillar, config and package states
-        generatePillar(minion);
+        MinionPillarManager.INSTANCE.generatePillar(minion);
         generateConfigState(newStateRev);
         StatesAPI.generateServerPackageState(minion);
     }
@@ -659,13 +411,6 @@ public enum SaltStateGeneratorService {
      */
     public void setSuseManagerStatesFilesRoot(Path generatedSlsRootIn) {
         this.suseManagerStatesFilesRoot = generatedSlsRootIn;
-    }
-
-    /**
-     * @param pillarDataPathIn the root path where pillar files are generated
-     */
-    public void setPillarDataPath(Path pillarDataPathIn) {
-        this.pillarDataPath = pillarDataPathIn;
     }
 
     /**
